### 2020-10-17 #62 @matbesancon
@simonbyrne you will need a "paper" folder at the top-level in the paper branch, where you can add all the paper-related files
