### 2020-10-17 #62 @simonbyrne
@matbesancon I tried that: https://github.com/JuliaParallel/MPI.jl/tree/paper/paper
