### 2020-10-19 #62 @matbesancon
You will need at least the YAML file, see https://github.com/JuliaCon/JuliaConSubmission.jl/tree/master/paper
If you want to build locally you will also need the rest of them
