### 2020-10-19 #62 @simonbyrne
Does it have to be .tex file, or can it be a .md file like JOSS?
