+++ src/collective.jl
@@ -17,9 +17,12 @@ function Barrier(comm::Comm)
 end
 
 """
-    Bcast!(buf[, count=length(buf)], root::Integer, comm::Comm)
+    Bcast!(buf, root::Integer, comm::Comm)
 
-Broadcast the first `count` elements of the buffer `buf` from `root` to all processes.
+Broadcast the buffer `buf` from `root` to all processes in `comm`.
+
+# See also
+- [`bcast`](@ref)
 
 # External links
 $(_doc_external("MPI_Bcast"))
@@ -30,18 +33,27 @@ function Bcast!(buf::Buffer, root::Integer, comm::Comm)
     @mpichk ccall((:MPI_Bcast, libmpi), Cint,
                   (MPIPtr, Cint, MPI_Datatype, Cint, MPI_Comm),
                   buf.data, buf.count, buf.datatype, root, comm)
-    buf.data
+    return buf.data
 end
 function Bcast!(data, root::Integer, comm::Comm)
     Bcast!(Buffer(data), root, comm)
 end
 
+"""
+    bcast(obj, root::Integer, comm::Comm)
+
+Broadcast the object `obj` from rank `root` to all processes on `comm`. This is able to handle arbitrary data.
+
+# See also
+
+- [`Bcast!`](@ref)
+"""
 function bcast(obj, root::Integer, comm::Comm)
     isroot = Comm_rank(comm) == root
     count = Ref{Cint}()
     if isroot
         buf = MPI.serialize(obj)
-        count[] = length()
+        count[] = length(buf)
     end
     Bcast!(count, root, comm)
     if !isroot
@@ -51,25 +63,27 @@ function bcast(obj, root::Integer, comm::Comm)
     if !isroot
         obj = MPI.deserialize(buf)
     end
-    obj
+    return obj
 end
 
 """
     Scatter!(sendbuf, recvbuf, root::Integer, comm::Comm)
 
 Splits the buffer `sendbuf` in the `root` process into `Comm_size(comm)` chunks,
-sending the `j`-th chunk to the process of rank `j` into the `recvbuf` buffer.
+sending the `j`-th chunk to the process of rank `j-1` into the `recvbuf` buffer.
 
-`sendbuf` on the root process should be a [`MultiBuffer`](@ref), or an `Array` if the
-`count` can be determined from `recvbuf`. On non-root processes it is ignored, and can be
-`nothing`.
+`sendbuf` on the root process should be a [`UBuffer`](@ref) (an `Array` can also be
+passed directly if the `count` can be determined from `recvbuf`). On non-root processes it
+is ignored, and `nothing` can be passed instead.
 
 `recvbuf` is a [`Buffer`](@ref) object, or any object for which `Buffer(recvbuf)` is
-defined. On the root process, it can also be `nothing`, in which case it is ignored (this
-corresponds the behaviour of `MPI_IN_PLACE` in `MPI_Scatter`). For example
+defined. On the root process, it can also be [`MPI.IN_PLACE`](@ref), in which case it is
+unmodified.
+
+For example:
 ```
 if root == MPI.Comm_rank(comm)
-    Scatter!(MultiBuffer(buf, count), nothing, root, comm)
+    Scatter!(UBuffer(buf, count), MPI.IN_PLACE, root, comm)
 else
     Scatter!(nothing, buf, root, comm)        
 end
@@ -82,7 +96,7 @@ end
 # External links
 $(_doc_external("MPI_Scatter"))
 """
[…]
