+++ src/buffers.jl
@@ -57,6 +57,27 @@ Additionally, certain sentinel values can be used, e.g. `MPI_IN_PLACE` or `MPI_B
 """
 MPIPtr
 
+struct InPlace
+end
+
+"""
+    IN_PLACE
+
+A sentinel value that can be passed as a buffer argument for certain collective operations
+to use the same buffer for send and receive operations.
+
+- [`Scatter!`](@ref) and [`Scatterv!`](@ref): can be used as the `recvbuf` argument on the
+  root process.
+
+- [`Gather!`](@ref) and [`Gatherv!`](@ref): can be used as the `sendbuf` argument on the
+  root process.
+
+- [`Allgather!`](@ref), [`Allgatherv!`](@ref), [`Alltoall!`](@ref) and
+  [`Alltoallv!`](@ref): can be used as the `sendbuf` argument on all processes.
+
+"""
+const IN_PLACE = InPlace()
+
 
 """
     MPI.Buffer
@@ -125,6 +146,9 @@ function Buffer(sub::SubArray{T,N,P,I,false}) where {T,N,P,I<:Tuple{Vararg{Union
     Buffer(parent(sub), Cint(1), datatype)
 end
 
+Buffer(::InPlace) = Buffer(MPI_IN_PLACE, 0, DATATYPE_NULL)
+Buffer(::Nothing) = Buffer(nothing, 0, DATATYPE_NULL)
+
 """
     Buffer_send(data)
 
@@ -133,32 +157,37 @@ Construct a [`Buffer`](@ref) object for a send operation from `data`, allowing c
 """
 Buffer_send(data) = isbits(data) ? Buffer(Ref(data)) : Buffer(data)
 Buffer_send(str::String) = Buffer(str, sizeof(str), MPI.CHAR)
+Buffer_send(::InPlace) = Buffer(InPlace())
+Buffer_send(::Nothing) = Buffer(nothing)
 
 
 
-const BUFFER_NULL = Buffer(C_NULL, 0, DATATYPE_NULL)
-
-const IN_PLACE = Buffer(MPI_IN_PLACE, 0, DATATYPE_NULL)
 
 
 
 """
-An MPI buffer for chunked collective communication.
+    MPI.UBuffer
+
+An MPI buffer for chunked collective communication, where all chunks are of uniform size.
 
 # Fields
 $(DocStringExtensions.FIELDS)
 
 # Usage
 
-    MultiBuffer(data, count::Integer, nchunks::Union{Nothing, Integer}, datatype::Datatype)
+    UBuffer(data, count::Integer, nchunks::Union{Nothing, Integer}, datatype::Datatype)
 
 Generic constructor.
 
-    MultiBuffer(data, count::Integer)
+    UBuffer(data, count::Integer)
 
-Construct a `Buffer` backed by `data`, where `count` is the number of elements in each chunk.
+Construct a `UBuffer` backed by `data`, where `count` is the number of elements in each chunk.
+
+# See also
+
+- [`VBuffer`](@ref): similar, but supports chunks of non-uniform sizes.
 """
-struct MultiBuffer{A}
+struct UBuffer{A}
     """A Julia object referencing a region of memory to be used for communication. It is
     required that the object can be `cconvert`ed to an [`MPIPtr`](@ref)."""
     data::A
@@ -173,27 +202,66 @@ struct MultiBuffer{A}
     """The [`MPI.Datatype`](@ref) stored in the buffer."""
     datatype::Datatype
 end
-MultiBuffer(data, count::Integer, nchunks::Integer, datatype::Datatype) =
-    MultiBuffer(data, Cint(count), Cint(nchunks), datatype)
[…]

[…]
