### 2020-10-20 18f2149
fix deprecations, get tests passing
