+++ src/deprecated.jl
@@ -106,35 +106,93 @@ end false
 
 @deprecate(Scatter!(sendbuf, recvbuf, count::Integer, root::Integer, comm::Comm),
            Scatter!(UBuffer(sendbuf, count), recvbuf, root, comm), false)
+@deprecate(Scatter!(sendbuf, recvbuf::AbstractArray, count::Integer, root::Integer, comm::Comm),
+           Scatter!(UBuffer(sendbuf, count), view(recvbuf, 1:count), root, comm), false)
+@deprecate(Scatter!(sendbuf, recvbuf::Nothing, count::Integer, root::Integer, comm::Comm),
+           Scatter!(UBuffer(sendbuf, count), MPI.IN_PLACE, root, comm), false)
+@deprecate(Scatter!(sendbuf::Nothing, recvbuf, count::Integer, root::Integer, comm::Comm),
+           Scatter!(nothing, recvbuf, root, comm), false)
+@deprecate(Scatter!(sendbuf::Nothing, recvbuf::AbstractArray, count::Integer, root::Integer, comm::Comm),
+           Scatter!(nothing, view(recvbuf, 1:count), root, comm), false)
+
 @deprecate(Scatter(sendbuf, count::Integer, root::Integer, comm::Comm),
-           Scatter!(sendbuf, similar(sendbuf, count), count, root, comm), false)
+           MPI.Comm_rank(comm) == root ?
+           Scatter!(UBuffer(sendbuf, count), similar(sendbuf, count), root, comm) :
+           Scatter!(nothing, similar(sendbuf, count), root, comm), false)
 
 @deprecate(Scatterv!(sendbuf, recvbuf, counts::Vector, root::Integer, comm::Comm),
-           Scatterv!(VBuffer(sendbuf, counts), recvbuf, root, comm), false)
+           MPI.Comm_rank(comm) == root ?
+           Scatterv!(VBuffer(sendbuf, counts), recvbuf, root, comm) :
+           Scatterv!(nothing, recvbuf, root, comm), false)
+@deprecate(Scatterv!(sendbuf, recvbuf::AbstractArray, counts::Vector, root::Integer, comm::Comm),
+           MPI.Comm_rank(comm) == root ?
+           Scatterv!(VBuffer(sendbuf, counts), view(recvbuf,1:counts[MPI.Comm_rank(comm)+1]), root, comm) :
+           Scatterv!(nothing, view(recvbuf,1:counts[MPI.Comm_rank(comm)+1]), root, comm), false)
+@deprecate(Scatterv!(sendbuf, recvbuf::Nothing, counts::Vector, root::Integer, comm::Comm),
+           MPI.Comm_rank(comm) == root ?
+           Scatterv!(VBuffer(sendbuf, counts), MPI.IN_PLACE, root, comm) :
+           Scatterv!(nothing, MPI.IN_PLACE, root, comm), false)
+@deprecate(Scatterv!(sendbuf::Nothing, recvbuf, counts::Vector, root::Integer, comm::Comm),
+           Scatterv!(nothing, recvbuf, root, comm), false)
+@deprecate(Scatterv!(sendbuf::Nothing, recvbuf::AbstractArray, counts::Vector, root::Integer, comm::Comm),
+           Scatterv!(nothing, view(recvbuf,1:counts[MPI.Comm_rank(comm)+1]), root, comm), false)
+
 @deprecate(Scatterv(sendbuf, counts::Vector, root::Integer, comm::Comm),
-           Scatterv!(VBuffer(sendbuf, counts), similar(sendbuf, counts[Comm_rank(comm) + 1]), root, comm), false)
+           MPI.Comm_rank(comm) == root ?
[…]
