### 2020-10-20 4a11972
remove old deprecations
