+++ src/deprecated.jl
@@ -1,105 +1,7 @@
 import Base: @deprecate
 
-## deprecated in v0.13
 
-## deprecated in v0.12
-
-@deprecate(Cart_get(comm, maxdims), Cart_get(comm), false)
-
-@deprecate(Cart_coords!(comm, rank, maxdims::Integer, coords),
-           Cart_coords!(comm, rank, coords), false)
-
-@deprecate(Cart_coords(comm, maxdims::Integer), Cart_coords(comm), false)
-
-
-## deprecated in v0.11
-
-@deprecate Reduce_in_place!(buf, count::Integer, op, root::Integer, comm::Comm) begin
-    @assert count == length(buf)
-    Reduce!(buf, op, root, comm)
-end false
-
-@deprecate Scatter_in_place!(buf, count::Integer, root::Integer, comm::Comm) begin
-    if root == MPI.Comm_rank(comm)
-        Scatter!(buf, nothing, count, root, comm)
-    else
-        Scatter!(nothing, buf, count, root, comm)        
-    end
-end false
-
-@deprecate Scatterv_in_place!(buf, counts::Vector, root::Integer, comm::Comm) begin
-    if root == MPI.Comm_rank(comm)
-        Scatterv!(buf, nothing, counts, root, comm)
-    else
-        Scatterv!(nothing, buf, counts, root, comm)
-    end
-end false
-
-@deprecate Gather_in_place!(buf, count::Integer, root::Integer, comm::Comm) begin
-    if root == MPI.Comm_rank(comm)
-        Gather!(nothing, buf, count, root, comm)
-    else
-        Gather!(buf, nothing, count, root, comm)
-    end
-end false    
-
-@deprecate Gatherv_in_place!(buf, counts::Vector{Cint}, root::Integer, comm::Comm) begin
-    if root == MPI.Comm_rank(comm)
-        Gatherv!(nothing, buf, counts, root, comm)
-    else
-        Gatherv!(buf, nothing, counts, root, comm)
-    end
-end false    
-
-@deprecate(mpitype(T), Datatype(T), false)
-
-@deprecate(Type_Create_Subarray(ndims::Integer, sizes::MPIBuffertype{Cint}, subsizes::MPIBuffertype{Cint},
-                                starts::MPIBuffertype{Cint}, order::Integer, oldtype),
-           Types.create_subarray(sizes, subsizes, starts, Datatype(oldtype); rowmajor = order == MPI_ORDER_C), false)
-@deprecate(Type_Create_Struct(nfields::Integer, blocklengths::MPIBuffertype{Cint},
-                              displacements::MPIBuffertype{Cptrdiff_t}, types::MPIBuffertype{MPI_Datatype}),
-           Types.create_struct(blocklengths, displacements, types), false)
-@deprecate(Type_Commit!(datatype), Types.commit!(datatype), false)
-
-
-@deprecate(Send(buf, count::Integer, datatype::Datatype, dest::Integer, tag::Integer, comm::Comm),
-           Send(Buffer(buf, count, datatype), dest, tag, comm), false)
-@deprecate(Send(buf::AbstractArray, count::Integer, dest::Integer, tag::Integer, comm::Comm),
-           Send(view(buf, 1:count), dest, tag, comm), false)
-@deprecate(Send(buf::Ref, count::Integer, dest::Integer, tag::Integer, comm::Comm),
-           Send(buf, dest, tag, comm), false)
-
-@deprecate(Isend(buf, count::Integer, datatype::Datatype, dest::Integer, tag::Integer, comm::Comm),
-           Isend(Buffer(buf,count,datatype), dest, tag, comm), false)
[…]
