+++ paper/header.tex
@@ -1,0 +1,19 @@
+% **************GENERATED FILE, DO NOT EDIT**************
+
+\title{My JuliaCon proceeding}
+
+\author[1]{1st author}
+\author[1, 2]{2nd author}
+\author[2]{3rd author}
+\affil[1]{University}
+\affil[2]{National Lab}
+
+\keywords{Julia, Optimization, Game theory, Compiler}
+
+\hypersetup{
+pdftitle = {My JuliaCon proceeding},
+pdfsubject = {JuliaCon 2019 Proceedings},
+pdfauthor = {1st author, 2nd author, 3rd author},
+pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+}
+
