+++ paper/paper.bib
@@ -25,7 +25,7 @@
 
 @misc{rsmpi,
   author = {{rsmpi developers}},
-  title = {{rsmpi}: MPI bindings for Rust},
+  title = {{rsmpi}: {MPI} bindings for {Rust}},
   year = {2020},
   publisher = {​GitHub},
 }
