### 2020-10-28 5355b49
Fixed small type in runtests.jl (#427)
