### 2020-11-03 10196d4
Add more combinations to GitHub actions
