+++ .github/workflows/UnitTests.yml
@@ -8,13 +8,15 @@ on:
       - master
 
 jobs:
-  test-os:
+  test-jll:
     timeout-minutes: 120
     strategy:
       matrix:
-        os: [macos-latest, windows-latest, ubuntu-18.04]
+        os: [macos-latest, windows-latest, ubuntu-latest]
         julia_version: ["1.3", "1.4", "1.5", "nightly"]
         include:
+          - os: macos-latest
+            mpiexec_args: "-host localhost"
           - os: macos-latest
             mpiexec_args: "-host localhost"
 
@@ -33,6 +35,55 @@ jobs:
 
     - name: Checkout
       uses: actions/checkout@v2.2.0
+
+    - uses: julia-actions/setup-julia@latest
+      with:
+        version: ${{ matrix.julia_version }}
+
+    # https://discourse.julialang.org/t/recommendation-cache-julia-artifacts-in-ci-services/35484
+    - name: Cache artifacts
+      uses: actions/cache@v1
+      env:
+        cache-name: cache-artifacts
+      with:
+        path: ~/.julia/artifacts
+        key: ${{ runner.os }}-test-${{ env.cache-name }}-${{ hashFiles('**/Project.toml') }}
+        restore-keys: |
+          ${{ runner.os }}-test-${{ env.cache-name }}-
+          ${{ runner.os }}-test-
+          ${{ runner.os }}-
+
+    - uses: julia-actions/julia-buildpkg@latest
+    - uses: julia-actions/julia-runtest@latest
+
+  test-homebrew:
+    timeout-minutes: 120
+    strategy:
+      matrix:
+        mpi: [mpich, openmpi]
+        julia_version: ["1.3", "1.4", "1.5", "nightly"]
+
+      fail-fast: false
+
+    runs-on: macos-latest
+
+    env:
+      JULIA_MPI_BINARY: system
+      JULIA_MPI_PATH: /usr/local
+
+    steps:
+    - name: Cancel Previous Runs
+      uses: styfle/cancel-workflow-action@0.4.0
+      with:
+        access_token: ${{ github.token }}
+
+    - name: Checkout
+      uses: actions/checkout@v2.2.0
+
+    - name: Install MPI via homebrew
+      run: brew install $MPI
+      env:
+        MPI: ${{ matrix.mpi }}
 
     - uses: julia-actions/setup-julia@latest
       with:
