+++ .github/workflows/UnitTests.yml
@@ -104,3 +104,53 @@ jobs:
 
     - uses: julia-actions/julia-buildpkg@latest
     - uses: julia-actions/julia-runtest@latest
+
+
+  test-apt:
+    timeout-minutes: 120
+    strategy:
+      matrix:
+        mpi: [mpich, openmpi]
+        julia_version: ["1.3", "1.4", "1.5", "nightly"]
+
+      fail-fast: false
+
+    runs-on: ubuntu-latest
+
+    env:
+      JULIA_MPI_BINARY: system
+
+    steps:
+    - name: Cancel Previous Runs
+      uses: styfle/cancel-workflow-action@0.4.0
+      with:
+        access_token: ${{ github.token }}
+
+    - name: Checkout
+      uses: actions/checkout@v2.2.0
+
+    - name: Install MPI via apt
+      run: sudo apt-get install $MPI
+      env:
+        MPI: ${{ matrix.mpi }}
+
+    - uses: julia-actions/setup-julia@latest
+      with:
+        version: ${{ matrix.julia_version }}
+
+    # https://discourse.julialang.org/t/recommendation-cache-julia-artifacts-in-ci-services/35484
+    - name: Cache artifacts
+      uses: actions/cache@v1
+      env:
+        cache-name: cache-artifacts
+      with:
+        path: ~/.julia/artifacts
+        key: ${{ runner.os }}-test-${{ env.cache-name }}-${{ hashFiles('**/Project.toml') }}
+        restore-keys: |
+          ${{ runner.os }}-test-${{ env.cache-name }}-
+          ${{ runner.os }}-test-
+          ${{ runner.os }}-
+
+    - uses: julia-actions/julia-buildpkg@latest
+    - uses: julia-actions/julia-runtest@latest
+    
