+++ .github/workflows/UnitTests.yml
@@ -56,7 +56,7 @@ jobs:
     - uses: julia-actions/julia-buildpkg@latest
     - uses: julia-actions/julia-runtest@latest
 
-  test-homebrew:
+  test-system-brew:
     timeout-minutes: 120
     strategy:
       matrix:
@@ -106,19 +106,20 @@ jobs:
     - uses: julia-actions/julia-runtest@latest
 
 
-  test-apt:
+  test-system-apt:
     timeout-minutes: 120
     strategy:
       matrix:
-        mpi: [mpich, openmpi]
+        mpi: [mpich, libopenmpi-dev]
         julia_version: ["1.3", "1.4", "1.5", "nightly"]
 
       fail-fast: false
 
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-20.04
 
     env:
       JULIA_MPI_BINARY: system
+      OMPI_MCA_btl_base_warn_component_unused: 0
 
     steps:
     - name: Cancel Previous Runs
@@ -153,4 +154,53 @@ jobs:
 
     - uses: julia-actions/julia-buildpkg@latest
     - uses: julia-actions/julia-runtest@latest
-    
+
+  test-system-msmpi:
+    timeout-minutes: 120
+    strategy:
+      matrix:
+        julia_version: ["1.3", "1.4", "1.5", "nightly"]
+
+      fail-fast: false
+
+    runs-on: windows-latest
+
+    env:
+      JULIA_MPI_BINARY: system
+
+    steps:
+    - name: Cancel Previous Runs
+      uses: styfle/cancel-workflow-action@0.4.0
+      with:
+        access_token: ${{ github.token }}
+
+    - name: Checkout
+      uses: actions/checkout@v2.2.0
+
+    - name: Download Microsoft MPI
+      run: (new-object net.webclient).DownloadFile("https://download.microsoft.com/download/A/E/0/AE002626-9D9D-448D-8197-1EA510E297CE/msmpisetup.exe", "MSMpiSetup.exe")
+      shell: powershell
+
+    - name: Install Microsoft MPI
+      run: MSMpiSetup.exe -unattend -minimal
+      shell: cmd
+
+    - uses: julia-actions/setup-julia@latest
+      with:
+        version: ${{ matrix.julia_version }}
+
+    # https://discourse.julialang.org/t/recommendation-cache-julia-artifacts-in-ci-services/35484
+    - name: Cache artifacts
+      uses: actions/cache@v1
+      env:
+        cache-name: cache-artifacts
+      with:
+        path: ~/.julia/artifacts
+        key: ${{ runner.os }}-test-${{ env.cache-name }}-${{ hashFiles('**/Project.toml') }}
+        restore-keys: |
+          ${{ runner.os }}-test-${{ env.cache-name }}-
+          ${{ runner.os }}-test-
+          ${{ runner.os }}-
+
+    - uses: julia-actions/julia-buildpkg@latest
+    - uses: julia-actions/julia-runtest@latest
