+++ .github/workflows/UnitTests.yml
@@ -14,9 +14,11 @@ jobs:
       matrix:
         os: [macos-latest, windows-latest, ubuntu-latest]
         julia_version: ["1.3", "1.4", "1.5", "nightly"]
+        julia-arch: [x64, x86]
+        exclude:
+          - os: macOS-latest
+            julia-arch: x86
         include:
-          - os: macos-latest
-            mpiexec_args: "-host localhost"
           - os: macos-latest
             mpiexec_args: "-host localhost"
 
@@ -61,7 +63,8 @@ jobs:
     strategy:
       matrix:
         mpi: [mpich, openmpi]
-        julia_version: ["1.3", "1.4", "1.5", "nightly"]
+        mpi_abi: ["", "unknown"]
+        julia_version: ["1.5", "nightly"]
 
       fail-fast: false
 
@@ -70,6 +73,7 @@ jobs:
     env:
       JULIA_MPI_BINARY: system
       JULIA_MPI_PATH: /usr/local
+      JULIA_MPI_ABI: ${{ matrix.mpi_abi }}
 
     steps:
     - name: Cancel Previous Runs
@@ -111,7 +115,8 @@ jobs:
     strategy:
       matrix:
         mpi: [mpich, libopenmpi-dev]
-        julia_version: ["1.3", "1.4", "1.5", "nightly"]
+        mpi_abi: ["", "unknown"]
+        julia_version: ["1.5", "nightly"]
 
       fail-fast: false
 
@@ -119,6 +124,7 @@ jobs:
 
     env:
       JULIA_MPI_BINARY: system
+      JULIA_MPI_ABI: ${{ matrix.mpi_abi }}
       OMPI_MCA_btl_base_warn_component_unused: 0
 
     steps:
@@ -159,7 +165,7 @@ jobs:
     timeout-minutes: 120
     strategy:
       matrix:
-        julia_version: ["1.3", "1.4", "1.5", "nightly"]
+        julia_version: ["1.5", "nightly"]
 
       fail-fast: false
 
@@ -179,11 +185,11 @@ jobs:
       uses: actions/checkout@v2.2.0
 
     - name: Download Microsoft MPI
-      run: (new-object net.webclient).DownloadFile("https://download.microsoft.com/download/A/E/0/AE002626-9D9D-448D-8197-1EA510E297CE/msmpisetup.exe", "MSMpiSetup.exe")
+      run: (new-object net.webclient).DownloadFile("https://download.microsoft.com/download/a/5/2/a5207ca5-1203-491a-8fb8-906fd68ae623/msmpisetup.exe", "msmpisetup.exe")
       shell: powershell
 
     - name: Install Microsoft MPI
-      run: MSMpiSetup.exe -unattend -minimal
+      run: msmpisetup.exe -unattend -minimal
       shell: cmd
 
     - uses: julia-actions/setup-julia@latest
