+++ .github/workflows/UnitTests.yml
@@ -14,10 +14,10 @@ jobs:
       matrix:
         os: [macos-latest, windows-latest, ubuntu-latest]
         julia_version: ["1.3", "1.4", "1.5", "nightly"]
-        julia-arch: [x64, x86]
+        julia_arch: [x64, x86]
         exclude:
           - os: macOS-latest
-            julia-arch: x86
+            julia_arch: x86
         include:
           - os: macos-latest
             mpiexec_args: "-host localhost"
@@ -40,6 +40,7 @@ jobs:
 
     - uses: julia-actions/setup-julia@latest
       with:
+        arch: ${{ matrix.julia_arch }}
         version: ${{ matrix.julia_version }}
 
     # https://discourse.julialang.org/t/recommendation-cache-julia-artifacts-in-ci-services/35484
