### 2020-11-04 d4b3d6c
disable Travis and Appveyor, update badges
