+++ .appveyor.yml
@@ -1,64 +1,0 @@
-environment:
-  matrix:
-  - julia_version: 1.3
-  - julia_version: 1.4
-  - julia_version: 1.4
-    mpi_impl: msmpi
-    JULIA_MPI_BINARY: system
-  - julia_version: nightly
-
-platform:
-  - x86 # 32-bit
-  - x64 # 64-bit
-
-matrix:
-  allow_failures:
-    - julia_version: nightly
-
-branches:
-  only:
-    - master
-    - /release-.*/
-
-notifications:
-  - provider: Email
-    on_build_success: false
-    on_build_failure: false
-    on_build_status_changed: false
-
-
-install:
-  - ps: iex ((new-object net.webclient).DownloadString("https://raw.githubusercontent.com/JuliaCI/Appveyor.jl/version-1/bin/install.ps1"))
-
-for:
--
-  matrix:
-    only:
-      - julia_version: 1.0
-        mpi_impl: msmpi
-      - julia_version: 1.4
-        mpi_impl: msmpi
-
-  install:
-    - ps: iex ((new-object net.webclient).DownloadString("https://raw.githubusercontent.com/JuliaCI/Appveyor.jl/version-1/bin/install.ps1"))
-    - ps: (new-object net.webclient).DownloadFile(
-          "https://download.microsoft.com/download/A/E/0/AE002626-9D9D-448D-8197-1EA510E297CE/msmpisetup.exe",
-          "C:\projects\MSMpiSetup.exe")
-    - C:\projects\MSMpiSetup.exe -unattend -minimal
-    # This shouldn't typicallybe needed: the installer adds the directory to
-    # the PATH. For some reason this doesn't work on Appveyor
-    - set JULIA_MPIEXEC=C:\Program Files\Microsoft MPI\Bin\mpiexec.exe
-
-build_script:
-  - echo "%JL_BUILD_SCRIPT%"
-  - C:\julia\bin\julia -e "%JL_BUILD_SCRIPT%"
-
-test_script:
-  - echo "%JL_TEST_SCRIPT%"
-  - C:\julia\bin\julia -e "%JL_TEST_SCRIPT%"
-
-# # Uncomment to support code coverage upload. Should only be enabled for packages
-# # which would have coverage gaps without running on Windows
-# on_success:
-#   - echo "%JL_CODECOV_SCRIPT%"
-#   - C:\julia\bin\julia -e "%JL_CODECOV_SCRIPT%"
