### 2020-11-04 88c52fe
Switch CI to GitHub actions (#428)
Adds more combinations to GitHub Actions
Disables Travis and Appveyor, update badges
