+++ .github/workflows/UnitTests.yml
@@ -8,13 +8,17 @@ on:
       - master
 
 jobs:
-  test-os:
+  test-jll:
     timeout-minutes: 120
     strategy:
       matrix:
+        os: [macos-latest, windows-latest, ubuntu-latest]
+        julia_version: ["1.3", "1.4", "1.5", "nightly"]
+        julia_arch: [x64, x86]
+        exclude:
+          - os: macOS-latest
+            julia_arch: x86
         include:
-          - os: ubuntu-latest
-          - os: windows-latest
           - os: macos-latest
             mpiexec_args: "-host localhost"
 
@@ -34,10 +38,10 @@ jobs:
     - name: Checkout
       uses: actions/checkout@v2.2.0
 
-    - name: Set up Julia
-      uses: julia-actions/setup-julia@v1.1.6
+    - uses: julia-actions/setup-julia@latest
       with:
-        version: 1.4
+        arch: ${{ matrix.julia_arch }}
+        version: ${{ matrix.julia_version }}
 
     # https://discourse.julialang.org/t/recommendation-cache-julia-artifacts-in-ci-services/35484
     - name: Cache artifacts
@@ -52,10 +56,159 @@ jobs:
           ${{ runner.os }}-test-
           ${{ runner.os }}-
 
-    - name: Install Project Packages
-      run: |
-        julia --project=@. -e 'using Pkg; Pkg.instantiate(); Pkg.build(verbose=true)'
+    - uses: julia-actions/julia-buildpkg@latest
+    - uses: julia-actions/julia-runtest@latest
 
-    - name: Run Unit Tests
-      run: |
-        julia --project=@. -e 'using Pkg; Pkg.test()'
+  test-system-brew:
+    timeout-minutes: 120
+    strategy:
+      matrix:
+        mpi: [mpich, openmpi]
+        mpi_abi: ["", "unknown"]
+        julia_version: ["1.5", "nightly"]
+
+      fail-fast: false
+
+    runs-on: macos-latest
+
+    env:
+      JULIA_MPI_BINARY: system
+      JULIA_MPI_PATH: /usr/local
+      JULIA_MPI_ABI: ${{ matrix.mpi_abi }}
+
+    steps:
+    - name: Cancel Previous Runs
+      uses: styfle/cancel-workflow-action@0.4.0
+      with:
+        access_token: ${{ github.token }}
+
+    - name: Checkout
+      uses: actions/checkout@v2.2.0
+
+    - name: Install MPI via homebrew
+      run: brew install $MPI
+      env:
+        MPI: ${{ matrix.mpi }}
+
+    - uses: julia-actions/setup-julia@latest
+      with:
+        version: ${{ matrix.julia_version }}
+
+    # https://discourse.julialang.org/t/recommendation-cache-julia-artifacts-in-ci-services/35484
+    - name: Cache artifacts
+      uses: actions/cache@v1
+      env:
+        cache-name: cache-artifacts
+      with:
+        path: ~/.julia/artifacts
+        key: ${{ runner.os }}-test-${{ env.cache-name }}-${{ hashFiles('**/Project.toml') }}
+        restore-keys: |
+          ${{ runner.os }}-test-${{ env.cache-name }}-
+          ${{ runner.os }}-test-
+          ${{ runner.os }}-
+
+    - uses: julia-actions/julia-buildpkg@latest
+    - uses: julia-actions/julia-runtest@latest
+
+
+  test-system-apt:
+    timeout-minutes: 120
+    strategy:
+      matrix:
+        mpi: [mpich, libopenmpi-dev]
+        mpi_abi: ["", "unknown"]
+        julia_version: ["1.5", "nightly"]
+
[…]
