### 2020-11-05 bf7aad1
Warn if wrong MPI launcher (#431)
Checks environment variables set by MPI launcher match values reported by MPI.
