### 2020-11-06 a0d00fc
add function to call init hooks (#433)
