### 2020-11-06 cc98f56
remove reference counting in lieu of Finalized() checks
