### 2020-11-06 33627b8
check finalized atexit
