### 2020-11-07 214065b
fix atexit hook
