### 2020-11-07 c5b1027
remove reference counting in lieu of Finalized() checks (#434)
* remove reference counting in lieu of Finalized() checks
* check finalized atexit
* update docs
