+++ src/environment.jl
@@ -25,43 +25,6 @@ function mpiexec(fn)
     _mpiexec(cmd -> fn(`$cmd $(Base.shell_split(get(ENV, "JULIA_MPIEXEC_ARGS", "")))`))
 end
 
-const REFCOUNT = Threads.Atomic{Int}(-1)
-
-"""
-    refcount_inc()
-
-Increment the MPI reference counter. This should be called at initialization of any object
-which calls an MPI routine in its finalizer. A matching [`refcount_dec`](@ref) should be
-added to the finalizer.
-
-For more details, see [Finalizers](@ref).
-"""
-function refcount_inc()
-    Threads.atomic_add!(REFCOUNT, 1)
-end
-
-"""
-    refcount_dec()
-
-Decrement the MPI reference counter. This should be added after an MPI call in an object
-finalizer, with a matching [`refcount_inc`](@ref) when the object is initialized.
-
-For more details, see [Finalizers](@ref).
-"""
-function refcount_dec()
-    # refcount zero, all objects finalized, now finalize MPI
-    if Threads.atomic_sub!(REFCOUNT, 1) == 1
-        if !Finalized()
-            # MPI can now be finalized, but MPI_Finalize is a collective and can act
-            # like a barrier (this may be implementation specific), if we are terminating
-            # due to a Julia exception, we should calling MPI_Finalize. We thus peek at the
-            # current exception, and only if that field is nothing do we terminate.
-            if ccall(:jl_current_exception, Any, ()) === nothing
-                _Finalize()
-            end
-        end
-    end
-end
 
 # Administrative functions
 function _warn_if_wrong_mpi()
@@ -88,11 +51,24 @@ function run_init_hooks()
     return nothing
 end
 
+function _finalize()
+    # MPI_Finalize is a collective and can act like a barrier (this may be implementation
+    # specific). If we are terminating due to a Julia exception, we shouldn't call
+    # MPI_Finalize. We thus peek at the current exception, and only if that field is
+    # nothing do we terminate.
+    if !Finalized() && ccall(:jl_current_exception, Any, ()) === nothing
+        Finalize()
+    end
+end
+
+
+
 
 """
-    Init()
+    Init(;finalize_atexit=true)
 
-Initialize MPI in the current process.
+Initialize MPI in the current process, and if `finalize_atexit` is true and adds an
+`atexit` hook to call [`MPI.Finalize`](@ref) if it hasn't already been called.
 
 All MPI programs must contain exactly one call to `MPI.Init` or
 [`MPI.Init_thread`](@ref). In particular, note that it is not valid to call `MPI.Init` or
@@ -104,11 +80,11 @@ The only MPI functions that may be called before `MPI.Init`/`MPI.Init_thread` ar
 # External links
 $(_doc_external("MPI_Init"))
 """
-function Init()
-    REFCOUNT[] == -1 || error("MPI.REFCOUNT in incorrect state: MPI may only be initialized once per session.")
+function Init(;finalize_atexit=true)
     @mpichk ccall((:MPI_Init, libmpi), Cint, (Ptr{Cint},Ptr{Cint}), C_NULL, C_NULL)
-    REFCOUNT[] = 1
-    atexit(refcount_dec)
+    if finalize_atexit
+        atexit(_finalize)
+    end
 
     run_init_hooks()
     _warn_if_wrong_mpi()
@@ -143,10 +119,12 @@ end
 
 
 """
-    Init_thread(required::ThreadLevel)
+    Init_thread(required::ThreadLevel; finalize_atexit=true)
 
-Initialize MPI and the MPI thread environment in the current process. The argument
[…]
