+++ src/collective.jl
@@ -67,30 +67,27 @@ function bcast(obj, root::Integer, comm::Comm)
 end
 
 """
-    Scatter!(sendbuf, recvbuf, root::Integer, comm::Comm)
+    Scatter!(sendbuf::Union{UBuffer,Nothing}, recvbuf, root::Integer, comm::Comm)
 
 Splits the buffer `sendbuf` in the `root` process into `Comm_size(comm)` chunks,
 sending the `j`-th chunk to the process of rank `j-1` into the `recvbuf` buffer.
 
 `sendbuf` on the root process should be a [`UBuffer`](@ref) (an `Array` can also be
-passed directly if the `count` can be determined from `recvbuf`). On non-root processes it
+passed directly if the sizes can be determined from `recvbuf`). On non-root processes it
 is ignored, and `nothing` can be passed instead.
 
 `recvbuf` is a [`Buffer`](@ref) object, or any object for which `Buffer(recvbuf)` is
 defined. On the root process, it can also be [`MPI.IN_PLACE`](@ref), in which case it is
-unmodified.
-
-For example:
+unmodified. For example:
 ```
 if root == MPI.Comm_rank(comm)
-    Scatter!(UBuffer(buf, count), MPI.IN_PLACE, root, comm)
+    MPI.Scatter!(UBuffer(buf, count), MPI.IN_PLACE, root, comm)
 else
-    Scatter!(nothing, buf, root, comm)        
+    MPI.Scatter!(nothing, buf, root, comm)        
 end
 ```
 
 # See also
-- [`Scatter`](@ref) to allocate the output buffer.
 - [`Scatterv!`](@ref) if the number of elements varies between processes.
 
 # External links
@@ -120,25 +117,27 @@ Scatter!(sendbuf::AbstractArray, recvbuf::Union{Ref,AbstractArray}, root::Intege
     Scatter!(UBuffer(sendbuf,length(recvbuf)), recvbuf, root, comm)    
 
 """
-    Scatterv!(sendbuf, recvbuf, counts, root, comm)
-
+    Scatterv!(sendbuf::Union{VBuffer,Nothing}, recvbuf, root, comm)
 
 Splits the buffer `sendbuf` in the `root` process into `Comm_size(comm)` chunks and sends
-the `j`th chunk to the process of rank `j-1` into the `recvbuf` buffer, which must be of
-length at least `count`.
+the `j`th chunk to the process of rank `j-1` into the `recvbuf` buffer.
 
-`recvbuf` can be [`MPI.IN_PLACE`](@ref) on the root process.
+`sendbuf` on the root process should be a [`VBuffer`](@ref). On non-root processes it is
+ignored, and `nothing` can be passed instead.
+
+`recvbuf` is a [`Buffer`](@ref) object, or any object for which `Buffer(recvbuf)` is
+defined. On the root process, it can also be [`MPI.IN_PLACE`](@ref), in which case it is
+unmodified. For example:
 ```
 if root == MPI.Comm_rank(comm)
-    Scatterv!(buf, nothing, counts, root, comm)
+    MPI.Scatterv!(VBuffer(buf, counts), MPI.IN_PLACE, root, comm)
 else
-    Scatterv!(nothing, buf, counts, root, comm)
+    MPI.Scatterv!(nothing, buf, root, comm)
 end
 ```
 
 # See also
-- [`Scatterv`](@ref) for the allocating operation
-- [`Scatter!`](@ref)/[`Scatter`](@ref) if the counts are the same for all processes
+- [`Scatter!`](@ref) if the number of elements are the same for all processes
 
 # External links
 $(_doc_external("MPI_Scatterv"))
[…]
