+++ docs/src/buffers.md
@@ -1,0 +1,13 @@
+# Buffers
+
+Buffers are used for sending and receiving data. MPI.jl provides the following buffer types:
+
+```@docs
+MPI.IN_PLACE
+MPI.Buffer
+MPI.Buffer_send
+MPI.UBuffer
+MPI.VBuffer
+MPI.RBuffer
+MPI.MPIPtr
+```
