### 2020-11-09 bfe4dd0
add Intel MPI to test matrix
