+++ .github/workflows/UnitTests.yml
@@ -162,6 +162,69 @@ jobs:
     - uses: julia-actions/julia-buildpkg@latest
     - uses: julia-actions/julia-runtest@latest
 
+
+  test-intel-linux:
+    timeout-minutes: 20
+    strategy:
+      matrix:
+        julia_version: ["1.5", "nightly"]
+
+      fail-fast: false
+
+    runs-on: ubuntu-20.04
+
+    env:
+      JULIA_MPI_BINARY: system
+
+    steps:
+    - name: Cancel Previous Runs
+      uses: styfle/cancel-workflow-action@0.4.0
+      with:
+        access_token: ${{ github.token }}
+
+    - name: Checkout
+      uses: actions/checkout@v2.2.0
+
+    - name: Install Intel MPI
+      run: |
+        wget http://registrationcenter-download.intel.com/akdlm/irc_nas/tec/16546/l_mpi_2019.7.217.tgz
+        tar -xzf l_mpi_2019.7.217.tgz
+        pushd l_mpi_2019.7.217
+        cat << EOF > intel.conf
+        ACCEPT_EULA=accept
+        CONTINUE_WITH_OPTIONAL_ERROR=yes
+        PSET_INSTALL_DIR=${HOME}/intel
+        CONTINUE_WITH_INSTALLDIR_OVERWRITE=no
+        PSET_MODE=install
+        ARCH_SELECTED=ALL
+        COMPONENTS=;intel-conda-index-tool__x86_64;intel-comp-l-all-vars__noarch;intel-comp-nomcu-vars__noarch;intel-imb__x86_64;intel-mpi-rt__x86_64;intel-mpi-sdk__x86_64;intel-mpi-doc__x86_64;intel-mpi-samples__x86_64;intel-mpi-installer-license__x86_64;intel-conda-impi_rt-linux-64-shadow-package__x86_64;intel-conda-impi-devel-linux-64-shadow-package__x86_64;intel-mpi-psxe__x86_64;intel-psxe-common__noarch;intel-psxe-common-doc__noarch;intel-compxe-pset
+        EOF
+        ./install.sh --silent intel.conf
+        popd
+        source ${HOME}/intel/compilers_and_libraries/linux/mpi/intel64/bin/mpivars.sh release
+
+    - uses: julia-actions/setup-julia@latest
+      with:
+        version: ${{ matrix.julia_version }}
+
+    # https://discourse.julialang.org/t/recommendation-cache-julia-artifacts-in-ci-services/35484
+    - name: Cache artifacts
+      uses: actions/cache@v1
+      env:
+        cache-name: cache-artifacts
+      with:
+        path: ~/.julia/artifacts
+        key: ${{ runner.os }}-test-${{ env.cache-name }}-${{ hashFiles('**/Project.toml') }}
+        restore-keys: |
+          ${{ runner.os }}-test-${{ env.cache-name }}-
+          ${{ runner.os }}-test-
+          ${{ runner.os }}-
+
+    - uses: julia-actions/julia-buildpkg@latest
+    - uses: julia-actions/julia-runtest@latest
+
+
+
   test-system-msmpi:
     timeout-minutes: 20
     strategy:
