+++ .github/workflows/UnitTests.yml
@@ -171,7 +171,7 @@ jobs:
 
       fail-fast: false
 
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-18.04  # 20.04 not supported
 
     env:
       JULIA_MPI_BINARY: system
