+++ .github/workflows/UnitTests.yml
@@ -222,7 +222,7 @@ jobs:
 
     - run: |
         source ${HOME}/intel/compilers_and_libraries/linux/mpi/intel64/bin/mpivars.sh release
-        julia --project -e 'using Pkg; Pkg.instantiate(verbose=true)'
+        julia --project -e 'using Pkg; Pkg.instantiate(verbose=true); Pkg.build(verbose=true)'
     - uses: julia-actions/julia-runtest@latest
 
 
