### 2020-11-09 ac09e05
load source in each step
