### 2020-11-09 6906525
add Intel MPI to test matrix (#435)
