+++ .github/workflows/UnitTests.yml
@@ -189,17 +189,17 @@ jobs:
       id: cache-intelmpi
       uses: actions/cache@v2
       with:
-        path: l_mpi_2019.7.217.tgz
-        key: ${{ runner.os }}-intelmpi-2019.7.217
+        path: l_mpi_2019.9.304.tgz
+        key: ${{ runner.os }}-intelmpi-2019.9.304
 
     - name: Download Intel MPI
       if: steps.cache-intelmpi.outputs.cache-hit != 'true'
-      run: wget --no-verbose http://registrationcenter-download.intel.com/akdlm/irc_nas/tec/16546/l_mpi_2019.7.217.tgz
+      run: wget https://registrationcenter-download.intel.com/akdlm/irc_nas/tec/17263/l_mpi_2019.9.304.tgz
 
     - name: Install Intel MPI
       run: |        
-        tar -xzf l_mpi_2019.7.217.tgz
-        pushd l_mpi_2019.7.217
+        tar -xzf l_mpi_2019.9.304.tgz
+        pushd l_mpi_2019.9.304
         cat << EOF > intel.conf
         ACCEPT_EULA=accept
         CONTINUE_WITH_OPTIONAL_ERROR=yes
@@ -211,7 +211,6 @@ jobs:
         EOF
         ./install.sh --silent intel.conf
         popd
-        cat ${HOME}/intel/compilers_and_libraries/linux/mpi/intel64/bin/mpivars.sh
 
     - uses: julia-actions/setup-julia@latest
       with:
