+++ .github/workflows/UnitTests.yml
@@ -204,10 +204,10 @@ jobs:
         ACCEPT_EULA=accept
         CONTINUE_WITH_OPTIONAL_ERROR=yes
         PSET_INSTALL_DIR=${HOME}/intel
-        CONTINUE_WITH_INSTALLDIR_OVERWRITE=no
+        CONTINUE_WITH_INSTALLDIR_OVERWRITE=yes
         PSET_MODE=install
         ARCH_SELECTED=ALL
-        COMPONENTS=;intel-conda-index-tool__x86_64;intel-comp-l-all-vars__noarch;intel-comp-nomcu-vars__noarch;intel-imb__x86_64;intel-mpi-rt__x86_64;intel-mpi-sdk__x86_64;intel-mpi-doc__x86_64;intel-mpi-samples__x86_64;intel-mpi-installer-license__x86_64;intel-conda-impi_rt-linux-64-shadow-package__x86_64;intel-conda-impi-devel-linux-64-shadow-package__x86_64;intel-mpi-psxe__x86_64;intel-psxe-common__noarch;intel-psxe-common-doc__noarch;intel-compxe-pset
+        COMPONENTS=;intel-conda-index-tool__x86_64;intel-comp-l-all-vars__noarch;intel-comp-nomcu-vars__noarch;intel-imb__x86_64;intel-mpi-rt__x86_64;intel-mpi-sdk__x86_64;intel-mpi-doc__x86_64;intel-mpi-samples__x86_64;intel-mpi-installer-license__x86_64;intel-conda-impi_rt-linux-64-shadow-package__x86_64;intel-conda-impi-devel-linux-64-shadow-package__x86_64;intel-mpi-psxe__x86_64;intel-psxe-common__noarch;intel-compxe-pset
         EOF
         ./install.sh --silent intel.conf
         popd
