### 2020-11-09 502717a
More Intel MPI tweaks (#436)
* New Intel MPI version
* fix script
