### 2020-11-10 e6d1ca8
update tests for deprecations
