+++ test/test_exscan.jl
@@ -27,7 +27,7 @@ for T in setdiff([Base.uniontypes(MPI.MPIDatatype)...], [Char, Int8, UInt8])
     B = MPI.Exscan(A, *, comm)
     @test B isa ArrayType{T}
 
-    MPI.Exscan!(A, length(A), *, comm)
+    MPI.Exscan!(A, *, comm)
     if rank > 0
         @test A == ArrayType{T}(fill(T(prodrank), 4))
     end
