+++ test/test_scatter.jl
@@ -18,29 +18,29 @@ isroot = rank == root
 
 for T in Base.uniontypes(MPI.MPIDatatype)
 
-    # Allocating version
     A = isroot ? ArrayType{T}(1:size) : ArrayType{T}(undef, 1)
-    B = MPI.Scatter(A, 1, root, comm)
-    @test B isa ArrayType{T}
-    @test Array(B)[1] == T(rank+1)
 
     # Non Allocating version
     B = ArrayType{T}(undef, 1)
-    MPI.Scatter!(A, B, 1, root, comm)
+    MPI.Scatter!(A, B, root, comm)
     @test Array(B)[1] == T(rank+1)
 
     # In place version
     B = isroot ? copy(A) : ArrayType{T}(undef, 1)
     if root == MPI.Comm_rank(comm)
-        MPI.Scatter!(B, nothing, 1, root, comm)
+        MPI.Scatter!(UBuffer(B,1), MPI.IN_PLACE, root, comm)
     else
-        MPI.Scatter!(nothing, B, 1, root, comm)
+        MPI.Scatter!(nothing, B, root, comm)
     end
     @test Array(B)[1] == T(rank+1)
 
     # Test throwing
-    B = ArrayType{T}(undef, 0)
-    @test_throws BoundsError MPI.Scatter!(A, B, 1, root, comm)
+    if isroot
+        B = ArrayType{T}(undef, 0)
+        @test_throws DivideError MPI.Scatter!(A, B, root, comm)
+        B = ArrayType{T}(undef, 8)
+        @test_throws AssertionError MPI.Scatter!(A, B, root, comm)
+    end
 end
 
 MPI.Finalize()
