### 2020-11-10 3411efc
Add buffer objects for collective operations (#335)
* Add ChunkBuffers for collective operations
* fix deprecations, get tests passing
* remove old deprecations
* update docs
* update tests for deprecations
