+++ src/buffers.jl
@@ -57,11 +57,42 @@ Additionally, certain sentinel values can be used, e.g. `MPI_IN_PLACE` or `MPI_B
 """
 MPIPtr
 
+# MPI_IN_PLACE
+
+struct InPlace
+end
+Base.cconvert(::Type{MPIPtr}, ::InPlace) = MPI_IN_PLACE
+
+
+"""
+    MPI.IN_PLACE
+
+A sentinel value that can be passed as a buffer argument for certain collective operations
+to use the same buffer for send and receive operations.
+
+- [`Scatter!`](@ref) and [`Scatterv!`](@ref): can be used as the `recvbuf` argument on the
+  root process.
+
+- [`Gather!`](@ref) and [`Gatherv!`](@ref): can be used as the `sendbuf` argument on the
+  root process.
+
+- [`Allgather!`](@ref), [`Allgatherv!`](@ref), [`Alltoall!`](@ref) and
+  [`Alltoallv!`](@ref): can be used as the `sendbuf` argument on all processes.
+
+- [`Reduce!`](@ref) (root only), [`Allreduce!`](@ref), [`Scan!`](@ref) and
+  [`Exscan!`](@ref): can be used as `sendbuf` argument.
+
+"""
+const IN_PLACE = InPlace()
+
+# TODO: MPI_BOTTOM
+
 
 """
     MPI.Buffer
 
-An MPI buffer for communication operations.
+An MPI buffer for communication with a single rank. It is used for point-to-point
+communication and some collective operations.
 
 # Fields
 $(DocStringExtensions.FIELDS)
@@ -82,6 +113,10 @@ and `datatype`. Methods are provided for
  - `CUDA.CuArray` if CUDA.jl is loaded
  - `SubArray`s of an `Array` or `CUDA.CuArray` where the layout is contiguous, sequential or
    blocked.
+
+# See also
+
+- [`Buffer_send`](@ref)
 
 """
 struct Buffer{A}
@@ -125,6 +160,9 @@ function Buffer(sub::SubArray{T,N,P,I,false}) where {T,N,P,I<:Tuple{Vararg{Union
     Buffer(parent(sub), Cint(1), datatype)
 end
 
+Buffer(::InPlace) = Buffer(IN_PLACE, 0, DATATYPE_NULL)
+Buffer(::Nothing) = Buffer(nothing, 0, DATATYPE_NULL)
+
 """
     Buffer_send(data)
 
@@ -133,7 +171,194 @@ Construct a [`Buffer`](@ref) object for a send operation from `data`, allowing c
 """
 Buffer_send(data) = isbits(data) ? Buffer(Ref(data)) : Buffer(data)
 Buffer_send(str::String) = Buffer(str, sizeof(str), MPI.CHAR)
+Buffer_send(::InPlace) = Buffer(InPlace())
+Buffer_send(::Nothing) = Buffer(nothing)
 
 
 
-const BUFFER_NULL = Buffer(C_NULL, 0, DATATYPE_NULL)
+
+
+
+"""
+    MPI.UBuffer
+
+An MPI buffer for chunked collective communication, where all chunks are of uniform size.
+
+# Fields
+$(DocStringExtensions.FIELDS)
+
+# Usage
+
+    UBuffer(data, count::Integer, nchunks::Union{Nothing, Integer}, datatype::Datatype)
+
+Generic constructor.
+
+    UBuffer(data, count::Integer)
+
+Construct a `UBuffer` backed by `data`, where `count` is the number of elements in each chunk.
+
+# See also
+
+- [`VBuffer`](@ref): similar, but supports chunks of non-uniform sizes.
+"""
+struct UBuffer{A}
[…]
