### 2020-11-13 169a0fa
remove references to deprecated collectives
