### 2020-11-14 5d6fae8
remove references to deprecated collectives (#442)
