### 2020-11-19 38c4b19
Remove unnecessary views
Fixes #445
