### 2020-11-19 925d21f
Improve CUDA-aware MPI support and add CI
- Synchronize task stream before passing CUDA pointer to MPI
- Enable GPU CI

Co-authored-by: Valentin Churavy <v.churavy@gmail.com>
