+++ test/runtests.jl
@@ -6,6 +6,8 @@ using MPIPreferences
 using DoubleFloats
 if get(ENV, "JULIA_MPI_TEST_ARRAYTYPE", "") == "CuArray"
     import CUDA
+    CUDA.version()
+    CUDA.precompile_runtime()
     ArrayType = CUDA.CuArray
 else
     ArrayType = Array
