### 2020-11-19 86518d7
Show CUDA version info
