### 2020-12-15 badd0b1
fix variables names and comments
