+++ docs/examples/05-job_schedule.jl
@@ -1,7 +1,7 @@
 # This example demonstrates a job scheduling through adding the
-# number 100 to every component of the vector data. The master
-# assigns one element to each slave to compute the operation.
-# When the worker is finished, the master sends another element
+# number 100 to every component of the vector data. The root
+# assigns one element to each worker to compute the operation.
+# When the worker is finished, the root sends another element
 # until each element is added 100
 # Inspired on
 # https://www.hpc.ntnu.no/ntnu-hpc-group/vilje/user-guide/software/mpi-and-mpi-io-training-tutorial/basic-mpi/job-queue
@@ -14,7 +14,7 @@ function job_queue(data,f)
     comm = MPI.COMM_WORLD
     rank = MPI.Comm_rank(comm)
     world_size = MPI.Comm_size(comm)
-    workers = world_size - 1
+    nworkers = world_size - 1
 
     root = 0
 
@@ -24,19 +24,19 @@ function job_queue(data,f)
     send_mesg = Array{T}(undef, 1)
     recv_mesg = Array{T}(undef, 1)
 
-    if rank == root # I am the master
+    if rank == root # I am root
 
         idx_recv = 0
         idx_sent = 1
 
         new_data = Array{T}(undef, N)
         # Array of workers requests
-        sreqs_workers = Array{MPI.Request}(undef,workers)
+        sreqs_workers = Array{MPI.Request}(undef,nworkers)
         # -1 = start, 0 = channel not available, 1 = channel available
-        status_workers = ones(workers).*-1
+        status_workers = ones(nworkers).*-1
         
-        # Send world_size messages to workers
-        for dst in 1:workers
+        # Send message to workers
+        for dst in 1:nworkers
             if idx_sent > N
                 break
             end
@@ -45,13 +45,13 @@ function job_queue(data,f)
             idx_sent += 1
             sreqs_workers[dst] = sreq
             status_workers[dst] = 0
-            print("Master: Sent number $(send_mesg[1]) to Worker $dst\n")
+            print("Root: Sent number $(send_mesg[1]) to Worker $dst\n")
         end
 
         # Send and receive messages until all elements are added
         while idx_recv != N
             # Check to see if there is an available message to receive
-            for dst in 1:workers
+            for dst in 1:nworkers
                 if status_workers[dst] == 0
                     (flag, status) = MPI.Test!(sreqs_workers[dst])
                     if flag
@@ -59,7 +59,7 @@ function job_queue(data,f)
                     end
                 end
             end
-            for dst in 1:workers
+            for dst in 1:nworkers
                 if status_workers[dst] == 1
                     ismessage, status = MPI.Iprobe(dst,dst+32, comm)
                     if ismessage
@@ -67,7 +67,7 @@ function job_queue(data,f)
                         MPI.Recv!(recv_mesg, dst, dst+32, comm)
                         idx_recv += 1
                         new_data[idx_recv] = recv_mesg[1]
-                        print("Master: Received number $(recv_mesg[1]) from Worker $dst\n")
+                        print("Root: Received number $(recv_mesg[1]) from Worker $dst\n")
                         if idx_sent <= N
                             send_mesg[1] = data[idx_sent]
                             # Sends new message
@@ -75,24 +75,24 @@ function job_queue(data,f)
                             idx_sent += 1
                             sreqs_workers[dst] = sreq
                             status_workers[dst] = 1
-                            print("Master: Sent number $(send_mesg[1]) to Worker $dst\n")
+                            print("Root: Sent number $(send_mesg[1]) to Worker $dst\n")
                         end
                     end
                 end
             end
         end
         
-        for dst in 1:workers
+        for dst in 1:nworkers
             # Termination message to worker
[…]
