### 2020-12-15 1ebbbd9
Add new example to docs (#449)
* add new job schedule example to docs

* add reference

* fix variables names and comments
