### 2020-12-31 ec29936
dlopen the mpi shared object and do not explicitly give library in ccall's
