+++ src/io.jl
@@ -13,7 +13,7 @@ function open(comm::Comm, filename::AbstractString, amode::Cint, info::Info)
     file = FileHandle()
     # int MPI_File_open(MPI_Comm comm, const char *filename, int amode,
     #                   MPI_Info info, MPI_File *fh)
-    @mpichk ccall((:MPI_File_open, libmpi), Cint,
+    @mpichk ccall(:MPI_File_open, Cint,
                   (MPI_Comm, Cstring, Cint, MPI_Info, Ptr{MPI_File}),
                   comm, filename, amode, info, file)
     finalizer(close, file)
@@ -63,7 +63,7 @@ end
 function close(file::FileHandle)
     if file.val != FILE_NULL.val && !Finalized()
         # int MPI_File_close(MPI_File *fh)
-        @mpichk ccall((:MPI_File_close, libmpi), Cint,
+        @mpichk ccall(:MPI_File_close, Cint,
                       (Ptr{MPI_File},), file)
     end
     return nothing
@@ -85,7 +85,7 @@ $(_doc_external("MPI_File_set_view"))
 function set_view!(file::FileHandle, disp::Integer, etype::Datatype, filetype::Datatype, datarep::AbstractString, info::Info)
     # int MPI_File_set_view(MPI_File fh, MPI_Offset disp, MPI_Datatype etype,
     #                       MPI_Datatype filetype, const char *datarep, MPI_Info info)
-    @mpichk ccall((:MPI_File_set_view, libmpi), Cint,
+    @mpichk ccall(:MPI_File_set_view, Cint,
                   (MPI_File, MPI_Offset,  MPI_Datatype, MPI_Datatype, Cstring, MPI_Info),
                   file, disp, etype, filetype, datarep, info)
     return file
@@ -109,7 +109,7 @@ function get_byte_offset(file::FileHandle, offset::Integer)
     r_displ = Ref{MPI_Offset}()
     # int MPI_File_get_byte_offset(MPI_File fh, MPI_Offset offset,
     #              MPI_Offset *disp)
-    @mpichk ccall((:MPI_File_get_byte_offset, libmpi), Cint,
+    @mpichk ccall(:MPI_File_get_byte_offset, Cint,
                   (MPI_File, MPI_Offset,  Ptr{MPI_Offset}),
                   file, offset, r_displ)
     r_displ[]
@@ -128,7 +128,7 @@ $(_doc_external("MPI_File_sync"))
 """
 function sync(file::FileHandle)
     # int MPI_File_sync(MPI_File fh)
-    @mpichk ccall((:MPI_File_sync, libmpi), Cint, (MPI_File,), file)
+    @mpichk ccall(:MPI_File_sync, Cint, (MPI_File,), file)
     return nothing
 end
 
@@ -151,7 +151,7 @@ function read!(file::FileHandle, buf::Buffer)
     stat_ref = Ref{Status}(MPI.STATUS_EMPTY)
     # int MPI_File_read(MPI_File fh, void *buf,
     #                   int count, MPI_Datatype datatype, MPI_Status *status)
-    @mpichk ccall((:MPI_File_read, libmpi), Cint,
+    @mpichk ccall(:MPI_File_read, Cint,
                   (MPI_File, MPIPtr, Cint, MPI_Datatype, Ptr{Status}),
                   file, buf.data, buf.count, buf.datatype, stat_ref)
     return stat_ref[]
@@ -176,7 +176,7 @@ function read_all!(file::FileHandle, buf::Buffer)
     stat_ref = Ref{Status}(MPI.STATUS_EMPTY)
     # int MPI_File_read_all(MPI_File fh, void *buf,
[…]
