+++ src/error.jl
@@ -12,7 +12,7 @@ function error_string(err::MPIError)
     len_ref = Ref{Cint}()
     str_buf = Vector{UInt8}(undef, MPI_MAX_ERROR_STRING)
     # int MPI_Error_string(int errorcode, char *string, int *resultlen)
-    @mpichk ccall((:MPI_Error_string, libmpi), Cint,
+    @mpichk ccall(:MPI_Error_string, Cint,
                   (Cint, Ptr{UInt8}, Ref{Cint}), err.code, str_buf, len_ref)
     resize!(str_buf, len_ref[])
     return String(str_buf)

[…]
