### 2021-01-04 10dcdca
use global symbols on Unix to support LD_PRELOAD interception
