### 2021-01-22 d2d37bf
Update implementations.jl
