### 2021-01-26 74f31f2
use global symbols on Unix to support LD_PRELOAD interception (#451)
* use global symbols on Unix to support LD_PRELOAD interception

* dlopen in deps

* move dlopen to src

* Update implementations.jl
