### 2021-01-28 #62 @crstnbr
@simonbyrne It seems that the issue is that you have a `paper.md` in your `paper/` directory. This conflicts with the `paper.tex` that we use for the Julia proceedings. Can you remove the `paper.md` file and ask whedon to generate the article proof again?
