+++ paper/paper.md
@@ -1,246 +1,0 @@
----
-title: 'MPI.jl: Julia bindings for the Message Passing Interface'
-tags:
-  - Julia
-  - MPI
-  - distributed computing
-authors:
-  - name: Simon Byrne
-    orcid: 0000-0001-8048-6810
-    affiliation: 1
-  - name: Lucas C. Wilcox
-    affiliation: 2
-  - name: Valentin Churavy
-    orcid: 0000-0002-9033-165X
-    affiliation: 3
-affiliations:
- - name: California Institute of Technology
-   index: 1
- - name: Naval Postgraduate School
-   index: 2
- - name: Massachusetts Institute of Technology
-   index: 3
-date: 21 September 2020
-bibliography: paper.bib
-
----
-
-
-# Summary
-
-MPI.jl is a Julia package for using the Message Passing Interface (MPI), a standardized and widely-supported communication interface for distributed computing, with multiple open source and proprietary implementions. It roughly follows the C MPI interface, with some additional conveniences afforded by the Julia language such as automatic handling of buffer lengths and datatypes.
-
-# Statement of Need
-
-Now over 25 years old, MPI is the stalwart of high-performance computing communication, supported on everything from single machines to billion-dollar supercomputers. Despite its age, it supports several models of communication, and significant engineering effort goes into optimizing performance and supporting the latest networking hardware. 
-
-Although Julia provides its own suite of distibuted computing tools via the Distributed standard library, it is based on a controller/worker model and is currently unable to leverage fast networking hardware such as InfiniBand, which limits its scalability to large problems. MPI.jl leverages the well-established and proven technology, including extensions such as the CUDA-aware interfaces for multi-GPU communication. It is being used by multiple Julia projects, including the CliMA Earth system modelling project [@clima].
-
-# Simple example and running MPI programs
-
-Most MPI programs utilise a single-program, multiple-data (SPMD) model where multiple processes all run the same program and communicate via messages, with their data determined by the process rank (a 0-based ordering of the processes).
-
-An example of this is is a simple "round-robin" communication pattern in which each process sends a message containing its rank to its next neighbor using non-blocking point-to-point operations:
-```julia
-# sendrecv.jl
-using MPI
-MPI.Init()
-
-comm = MPI.COMM_WORLD
-rank = MPI.Comm_rank(comm)
-N = MPI.Comm_size(comm)
-
-# non-blocking receive from previous rank
-recv_buf = Array{Float64}(undef, 2)
-recv_req = MPI.Irecv!(recv_buf, mod(rank-1, N), 0, comm)
-
-# non-blocking send to next rank
-send_buf = Float64[rank, rank]
-send_req = MPI.Isend(send_buf, mod(rank+1, N), 0, comm)
-
-# block until communication is completed
-MPI.Waitall!([recv_req, send_req])
-print("$rank: Received $recv_buf\n")
-```
-
-This can be run using the MPI launcher (typically called `mpiexec`)
-```
-$ mpiexec -n 3 julia sendrecv.jl
-0: Received [2.0, 2.0]
-2: Received [1.0, 1.0]
-1: Received [0.0, 0.0]
-```
-
-# Implementation details and challenges
-
[…]
