### 2021-02-04 #62 @crstnbr
@simonbyrne Given that you submitted some time ago (sorry for the delay) do you want to make additional changes to the manuscript or the repo? If not, I would let Whedon close this thread and move us to review stage.
