### 2021-02-07 #68 @whedon
[REVIEW]: MPI.jl: Julia bindings for the Message Passing Interface
**Submitting author:** @simonbyrne (<a href="http://orcid.org/0000-0001-8048-6810">Simon Byrne</a>)
**Repository:** <a href="https://github.com/JuliaParallel/MPI.jl" target ="_blank">https://github.com/JuliaParallel/MPI.jl</a>
**Version:**
