### 2021-02-21 #68 @dalcinl
A few minor comments/requests for the article:

* Section 3.2, about `MPI.jl` automatically building and commiting datatypes. What about user-defined datatypes? Are those supported? Can those be used in communication calls?. Any plan to support that? I think this should be clarified in the text, even if just with a short sentence. `MPI.jl`'s datatype automatism looks even better than `mpi4py`'s, but support for a more low level, MPI-native way could also be of use for experienced users and library developers.

* Page 2, section 3.4, last sentence in right column before Sec. 4 starts: *how to deal with binaries which depend on MPI*. I'm not sure what this really means. Are you talking about interoperating with shared libraries containing MPI-based parallel algorithms? I would suggest a slight reword to make meaning of this sentence clearer.

* Page 3, left column, second paragraph, there is a missing blank space between `MPI.send` and the word "and".

* Page 3, left column, second paragraph. You blame Julia's GC for the slowdown with medium message sizes. Did you try to bracket the communication+timing with `gc_disable()`/`gc_enable()` to confirm, if that is ever possible? BTW, could you clarify what Python and NumPy versions were used in these tests? FYI, mpi4py branch master, to be released soon, now uses the highest available pickle protocol to communicate generic objects. Under Python 3.8, that would be pickle protocol version 5, which introduces important optimizations for buffer-like objects like NumPy arrays.
