### 2021-02-24 #68 @simonbyrne
>     * Some of the references do not have a DOI.

@crstnbr I've included the DOIs in the bib file: any idea why they don't appear?
