### 2021-02-24 #68 @stevengj
Some comments:

• why is mpi4py faster for large generic objects? Is it Python pickle vs Julia Serialization module?

• An important part of MPI is non-blocking communications, but asynchonous I/O is usually handled in julia with tasks or "green threads".  Are there any prospects to combine these? (recently asked on discourse)

• MPI persistent communications are not supported?  any plans?

• In section 3.2 what about non-Array containers supporting stride and pointer? Should there be discussion of the MPI.Buffer and MPI.UBuffer types and the design decisions here?

• dynamic process management in MPI-2 is not supported?  any plans?

• one-sided communications seem to be supported in MPI.jl but not are mentioned in the paper?

• parallel I/O support?  is it possible to use MPI.jl with parallel HDF5 and other libraries?

• The MPI.jl repo was started in 2012, and is one of the oldest Julia packages.  What have been the major milestones and changes over the last 8 years?   Any design mistakes or lessons learned?

typo: "arbirary" in sec. 3.1
