+++ paper/paper.bib
@@ -59,3 +59,19 @@ doi = {10.1002/2017GL076101},
 year = {2017}
 }
 
+@INPROCEEDINGS {spack,
+author = {T. Gamblin and M. LeGendre and M. R. Collette and G. L. Lee and A. Moody and B. R. de Supinski and S. Futral},
+booktitle = {SC15: International Conference for High-Performance Computing, Networking, Storage and Analysis},
+title = {The Spack package manager: bringing order to HPC software chaos},
+year = {2015},
+volume = {},
+issn = {2167-4337},
+pages = {1-12},
+keywords = {libraries;complexity theory;syntactics;operating systems;production},
+doi = {10.1145/2807591.2807623},
+url = {https://doi.ieeecomputersociety.org/10.1145/2807591.2807623},
+publisher = {IEEE Computer Society},
+address = {Los Alamitos, CA, USA},
+month = {nov}
+}
+
