+++ paper/paper.tex
@@ -37,6 +37,21 @@ extensions such as the CUDA-aware interfaces for multi-GPU
 communication. It is being used by multiple Julia projects, including
 the CliMA Earth system modelling project \cite{clima}.
 
+\subsection{History}
+
+MPI.jl originated as a repository developed by Lucas Wilcox in July 2012, 4 months after
+Julia was first publicly announced. It was registered in the Julia package registry in
+2014, and since then has evolved considerably along with the Julia language itself. It has
+so far received contributions from 51 people: many of these are users who required the
+addition of a specific piece of functionality.
+
+Major changes to the codebase were typically precipitated by changes in the Julia language
+itself, modifications to the build system to support different MPI implementations (see
+section \ref{sec:application-binary-interface}) or addition of support for new
+platforms. The interface has stayed remarkably consistent over time, the only major change
+at the highest level being the recent addition of buffer types (section
+\ref{sec:buffers-datatypes-and-operators}).
+
 \section{Simple example and running MPI programs}
 \label{sec:simple-example}
 
@@ -45,9 +60,9 @@ where multiple processes all run the same program and communicate via
 messages, with their data determined by the process rank (a 0-based
 ordering of the processes).
 
-An example of this is is a simple ``round-robin'' communication pattern
+An example of this is a simple ``round-robin'' communication pattern
 in which each process sends a message containing its rank to its next
-neighbor using non-blocking point-to-point operations:
+neighbor using nonblocking point-to-point operations:
 
 \begin{lstlisting}[language = Julia]
 # sendrecv.jl
@@ -59,12 +74,12 @@ comm = MPI.COMM_WORLD
 rank = MPI.Comm_rank(comm)
 N = MPI.Comm_size(comm)
 
-# non-blocking receive from previous rank
+# nonblocking receive from previous rank
 recv_buf = Array{Float64}(undef, 2)
 recv_req = MPI.Irecv!(recv_buf, mod(rank-1, N), 0,
                       comm)
 
-# non-blocking send to next rank
+# nonblocking send to next rank
 send_buf = Float64[rank, rank]
 send_req = MPI.Isend(send_buf, mod(rank+1, N), 0,
                      comm)
@@ -87,15 +102,31 @@ $ mpiexec -n 3 julia sendrecv.jl
 \section{Implementation details and challenges}
 \label{sec:implementation-details-and-challenges}
 
-Although MPI.jl mirrors the C MPI interface quite closely, it does take
-advantage of several features of the Julia language to improve
-usability. The C and Fortran MPI interfaces require that users manually
+Although MPI.jl mirrors the C MPI interface quite closely, it does take advantage of
+several features of the Julia language to improve usability. Many of these changes were
+heavily influenced by the mpi4py package \cite{dalcin2011parallel} which provides MPI
+bindings for Python, and has a similar aim of making MPI functionality easily useable in a
+high-level, dynamic language.
+
+The C and Fortran MPI interfaces require that users manually
 check the error code returned by each function; MPI.jl is able to use
 Julia's exception handling machinery to automatically check error codes
 and print readable error messages. This allows functions to return their
 results via return values instead of via additional functions arguments.
-For example, non-blocking operations return \verb Request  objects;
+For example, nonblocking operations return \verb Request  objects;
 blocking receive operations return their output buffers.
+
+\subsection{Functionality}
+
+Most commonly-used MPI functionality is currently available via the package, including
+point-to-point (blocking and nonblocking), collective, and one-sided communication
+operations as well as MPI I/O operations.
+
[…]
