### 2021-02-24 4c42158
Define wait(req) to use threadcall
