### 2021-02-25 9a5b6f4
benchmark serialize, use random numbers
