+++ benchmarks/Project.toml
@@ -2,3 +2,4 @@
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Serialization = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
