+++ benchmarks/pingpong.c
@@ -2,10 +2,11 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-srand((unsigned) time(&t));
-
 
 double pingpong(int bufsize, int iters) {
+
+  srand(7);
+  
   double *buffer = (double *)malloc(bufsize * sizeof(double));
   for (int i=0; i<bufsize; i++) {
     buffer[i] = rand();
