### 2021-02-27 768621a
Use the serial comma everywhere
Before we were sometimes using it and sometimes not.  Hopefully this
fixes that.
