### 2021-03-01 #68 @simonbyrne
Thank you @dalcinl and @stevengj for the feedback. We have addressed the comments as follows:

>   * Section 3.2, about `MPI.jl` automatically building and commiting datatypes. What about user-defined datatypes? Are those supported? Can those be used in communication calls?. Any plan to support that? I think this should be clarified in the text, even if just with a short sentence. `MPI.jl`'s datatype automatism looks even better than `mpi4py`'s, but support for a more low level, MPI-native way could also be of use for experienced users and library developers.

Yes, this available, we have added a sentence to this effect. (note it is now section 3.3)

>   * Page 2, section 3.4, last sentence in right column before Sec. 4 starts: _how to deal with binaries which depend on MPI_. I'm not sure what this really means. Are you talking about interoperating with shared libraries containing MPI-based parallel algorithms? I would suggest a slight reword to make meaning of this sentence clearer.

Yes, that is exactly the problem. We have expanded the text here, and given examples (HDF5 and PETSc).

>   * Page 3, left column, second paragraph, there is a missing blank space between `MPI.send` and the word "and".

Fixed.

>   * Page 3, left column, second paragraph. You blame Julia's GC for the slowdown with medium message sizes. Did you try to bracket the communication+timing with `gc_disable()`/`gc_enable()` to confirm, if that is ever possible? BTW, could you clarify what Python and NumPy versions were used in these tests? FYI, mpi4py branch master, to be released soon, now uses the highest available pickle protocol to communicate generic objects. Under Python 3.8, that would be pickle protocol version 5, which introduces important optimizations for buffer-like objects like NumPy arrays.

We did try disabling the GC, and it didn't have much effect. Instead we benchmarked the serialization protocols directly, which do appear to explain the main differences. We have clarified this and added the version numbers as well.

>   * Some of the references do not have a DOI.

This appears to be a problem with the style sheet.

>   * Perhaps you should add a disclaimer to the paper (similar to the one you have in your docs), that the project does not support the full feature set of the MPI standard, together with a promise that they this will be addressed in the future?

We have tried to address this in the new section 3.1.

> • why is mpi4py faster for large generic objects? Is it Python pickle vs Julia Serialization module?

This does appear to be the case. We have added comments to this effect in section 4.1.

> • An important part of MPI is non-blocking communications, but asynchonous I/O is usually handled in julia with tasks or "green threads". Are there any prospects to combine these? (recently asked on discourse)

We have added section 3.6 to address this.

> • MPI persistent communications are not supported? any plans?

They are not currently supported, we have mentioned it in section 3.1.

> • In section 3.2 what about non-Array containers supporting stride and pointer? Should there be discussion of the MPI.Buffer and MPI.UBuffer types and the design decisions here?

`UBuffer` was added since this paper was originally written. We have added it to (now) section 3.3.

> • dynamic process management in MPI-2 is not supported? any plans?

There is some support (mentioned in new section 3.1), but unfortunately their utility is hampered by the requirement to use within the MPI launcher (discussed at end of the new section 3.6).

> • one-sided communications seem to be supported in MPI.jl but not are mentioned in the paper?

It is mentioned in the new section 3.1.

> • parallel I/O support? is it possible to use MPI.jl with parallel HDF5 and other libraries?

I/O support is mentioned in new section 3.1, added discussion of integration with HDF5 in section 3.5.

> • The MPI.jl repo was started in 2012, and is one of the oldest Julia packages. What have been the major milestones and changes over the last 8 years? Any design mistakes or lessons learned?

We added a brief history in new section 1.1

In addition to the above changes, we added some context to the pooled variance example (section 4.3).
