### 2021-03-02 #68 @simonbyrne
I also added the doi links to the bibliography
