### 2021-03-11 #68 @dalcinl
@simonbyrne Great job, your clarifications improved the text considerably. A few minor issues remains:
- In the last line of section 3.1 you wrote "shold". I'm not a native English speaker, and after a quick search I found that word is an obsolete preterits of "shall". But even then, it does not look like your sentence is using a past tense. Is this just a typo, and you actually meant "should" ?
- Section 33, page 2, right column, near the top: You wrote "bitstype" two times. However, in [Julia docs](https://docs.julialang.org/en/v1/manual/types/) they use "bits type" with a space.
- Section 3.4, first line: please fix typo "appilication" -> "application". Just in case, please give it a final pass with a spell checker.

@crstnbr After Simon fixes these minor typos above, I'm done with this review, I consider the paper ready to go. Is this confirmation enough, or should I somehow tell whedon about it?
