### 2021-03-15 3aac247
Avoid InitError on Julia 1.3 + Unix + system MPI
