+++ src/implementations.jl
@@ -10,7 +10,8 @@ macro mpicall(expr)
 
     # On unix systems we call the global symbols to allow for LD_PRELOAD interception
     # It can be emulated in Windows (via Libdl.dllist), but this is not fast.
-    if Sys.isunix() && expr.args[2].head == :tuple
+    if Sys.isunix() && expr.args[2].head == :tuple &&
+            expr.args[2].args[1] != :(:MPI_Get_library_version)
         expr.args[2] = expr.args[2].args[1]
     end
 
