### 2021-03-18 1655da9
Avoid InitError on Julia 1.3 with system MPI libraries (#456)
* Avoid InitError on Julia 1.3 + Unix + system MPI
* Add version check
