### 2021-03-18 2cb6039
add basic error handler functionality
