### 2021-03-23 cf1a818
Allow `Waitall!` for an empty input `reqs = []`
