+++ src/pointtopoint.jl
@@ -447,7 +447,7 @@ $(_doc_external("MPI_Waitall"))
 """
 function Waitall!(reqs::Vector{Request})
     count = length(reqs)
-    count == 0 && return []
+    count == 0 && return Status[]
     reqvals = [reqs[i].val for i in 1:count]
     stats = fill(STATUS_EMPTY, count)
     # int MPI_Waitall(int count, MPI_Request array_of_requests[],
