### 2021-03-23 ccdb676
Update test/test_sendrecv.jl
Co-authored-by: Simon Byrne <simonbyrne@gmail.com>
