### 2021-03-24 dd2c587
Allow `Waitall!` for an empty input `reqs = []` (#458)
* Allow `Waitall!` for an empty input `reqs = []`
* Add a test
