### 2021-04-04 ea37f67
Add finalizer to `create_struct`
