### 2021-04-08 ce8e7b8
Define new method for `Datatype` to fix issue #462
