### 2021-04-08 f83140b
add basic error handler functionality (#457)
* add basic error handler functionality
* set default error handler to ERRORS_RETURN
