### 2021-04-09 ac94594
add missing methods
