+++ src/deprecated.jl
@@ -169,9 +169,17 @@ import Base: @deprecate
 
 @deprecate(Get(origin_buffer, count::Integer, target_rank::Integer, target_disp::Integer, win::Win),
            Get(view(origin_buffer, 1:count), target_rank, target_disp, win), false)
+@deprecate(Get(origin_buffer::Ref, count::Integer, target_rank::Integer, target_disp::Integer, win::Win),
+           Get(origin_buffer, target_rank, target_disp, win), false)
 @deprecate(Put(origin_buffer, count::Integer, target_rank::Integer, target_disp::Integer, win::Win),
            Put(view(origin_buffer, 1:count), target_rank, target_disp, win), false)
+@deprecate(Put(origin_buffer::Ref, count::Integer, target_rank::Integer, target_disp::Integer, win::Win),
+           Put(origin_buffer, target_rank, target_disp, win), false)
 @deprecate(Accumulate(origin_buffer, count::Integer, target_rank::Integer, target_disp::Integer, op::Op, win::Win),
            Accumulate(view(origin_buffer, 1:count), target_rank, target_disp, op, win), false)
+@deprecate(Accumulate(origin_buffer::Ref, count::Integer, target_rank::Integer, target_disp::Integer, op::Op, win::Win),
+           Accumulate(origin_buffer, target_rank, target_disp, op, win), false)
 @deprecate(Get_accumulate(origin_buffer, result_buffer, count::Integer, target_rank::Integer, target_disp::Integer, op::Op, win::Win),
            Get_accumulate(view(origin_buffer,1:count), view(result_buffer,1:count), target_rank, target_disp, op, win), false)
+@deprecate(Get_accumulate(origin_buffer::Ref, result_buffer::Ref, count::Integer, target_rank::Integer, target_disp::Integer, op::Op, win::Win),
+           Get_accumulate(origin_buffer, result_buffer, target_rank, target_disp, op, win), false)
