### 2021-04-15 b7d005f
docstrings for one-sided operations
