### 2021-04-20 4f917ba
loosen type restrictions on one-sided buffers
