### 2021-04-22 637c90c
add docs for one-sided functions
