+++ docs/src/onesided.md
@@ -4,4 +4,8 @@
 MPI.Win_create
 MPI.Win_create_dynamic
 MPI.Win_allocate_shared
+MPI.Get
+MPI.Put
+MPI.Accumulate
+MPI.Get_accumulate
 ```
