### 2021-04-22 d2cfb9f
Apply suggestions from code review
Co-authored-by: Simon Byrne <simonbyrne@gmail.com>
