### 2021-04-22 35601bf
amend docstring of `Buffer`
