### 2021-04-22 96070b7
docstrings for one-sided operations (#466)
* docstrings for one-sided operations
* amend docstring of `Buffer`
