### 2021-05-02 11b1335
Add Scatterv example based on #469
