### 2021-05-02 aca992f
improve split_count function
