### 2021-05-03 a584b85
Add Scatterv example based on #469 (#470)
