### 2021-05-10 76ef3c7
Improve Datatype handling
- Uses generated functions to avoid creating new instances `Datatype`s on every communication operation (#462). These are lazily initialized to allow for use in precompiled modules.
- Attach the corresponding Julia to an MPI Datatype as an attribute
- Nicer printing of Datatypes
- Allow specifying bits type as return type of MPI.Scatter
- Minor optimisation when constructing buffers of a single NTuple
