+++ src/datatypes.jl
@@ -5,18 +5,21 @@ A `Datatype` represents the layout of the data in memory.
 
 # Usage
 
-    Datatype(T; commit=true)
+    Datatype(T)
 
-Either return the predefined `Datatype` or create a new `Datatype` for the Julia type
-`T`. If `commit=true`, then the [`Types.commit!`](@ref) operation will also be applied so
-that it can be used for communication operations.
+Either return the predefined `Datatype` corresponding to `T`, or create a new `Datatype`
+for the Julia type `T`, calling [`Types.commit!`](@ref) so that it can be used for
+communication operations.
 
 Note that this can only be called on types for which `isbitstype(T)` is `true`.
 """
 @mpi_handle Datatype MPI_Datatype
 
 const DATATYPE_NULL = _Datatype(MPI_DATATYPE_NULL)
-Datatype() = Datatype(DATATYPE_NULL.val)
+
+const MPI_Datatype_default = MPI_Datatype == Cint ? MPI_DATATYPE_NULL : C_NULL
+Datatype() = Datatype(MPI_Datatype_default)
+    
 
 function free(dt::Datatype)
     if dt.val != DATATYPE_NULL.val && !Finalized()
@@ -25,6 +28,86 @@ function free(dt::Datatype)
     return nothing
 end
 
+# attributes
+function _type_null_copy_fn(oldtype::MPI_Datatype, type_keyval::Cint,
+                            extra_state::Ptr{Cvoid},
+                            attribute_val_in::Ptr{Cvoid}, attribute_val_out::Ptr{Cvoid},
+                            flag::Ptr{Cint})
+    unsafe_store!(flag, Cint(0))
+    return MPI_SUCCESS
+end
+function _type_null_delete_fn(oldtype::MPI_Datatype, type_keyval::Cint,
+                           attribute_val_in::Ptr{Cvoid}, extra_state::Ptr{Cvoid})
+    return MPI_SUCCESS
+end
+
+function create_keyval(::Type{Datatype})
+    ref = Ref(Cint(0))
+    @mpichk ccall((:MPI_Type_create_keyval, libmpi), Cint,
+                  (Ptr{Cvoid},
+                   Ptr{Cvoid},
+                   Ptr{Cint},
+                   Ptr{Cvoid}),
+                  @cfunction(_type_null_copy_fn,Cint,(MPI_Datatype, Cint, Ptr{Cvoid},
+                                                      Ptr{Cvoid}, Ptr{Cvoid}, Ptr{Cint})),
+                  @cfunction(_type_null_delete_fn,Cint,(MPI_Datatype, Cint, Ptr{Cvoid},
+                                                      Ptr{Cvoid})),
+                  ref, C_NULL)
+    return ref[]
+end
+function set_attr!(datatype::Datatype, keyval::Cint, attrval::Ptr{Cvoid})
+    @mpichk ccall((:MPI_Type_set_attr, libmpi), Cint,
+                  (MPI_Datatype, Cint, Ptr{Cvoid}), datatype, keyval, attrval)
+    return nothing
+end
+function get_attr(datatype::Datatype, keyval::Cint)
+    flagref = Ref(Cint(0))
+    attrref = Ref{Ptr{Cvoid}}(C_NULL)
+    @mpichk ccall((:MPI_Type_get_attr, libmpi), Cint,
+                  (MPI_Datatype, Cint, Ptr{Ptr{Cvoid}}, Ptr{Cint}), datatype, keyval, attrref, flagref)
+    if flagref[] == 0
+        return nothing
+    end
+    return attrref[]
+end
+function del_attr!(datatype::Datatype, keyval::Cint)
[…]
