### 2021-05-10 6eb22a0
duplicate Datatypes if same size as existing type
