### 2021-05-14 eeed089
Improve Datatype handling (#471)
* Improve Datatype handling

- Uses generated functions to avoid creating new instances `Datatype`s on every communication operation (#462). These are lazily initialized to allow for use in precompiled modules.
- Attach the corresponding Julia to an MPI Datatype as an attribute
- Nicer printing of Datatypes
- Allow specifying bits type as return type of MPI.Scatter
- Minor optimisation when constructing buffers of a single NTuple

* duplicate Datatypes if same size as existing type
