### 2021-06-05 ea8fb90
Add a bunch of barriers to io_shared tests (#474)
