+++ test/test_io_shared.jl
@@ -21,6 +21,8 @@ MPI.Barrier(comm)
 fh = MPI.File.open(comm, filename, read=true, write=true, create=true)
 @test MPI.File.get_position_shared(fh) == 0
 
+MPI.Barrier(comm)
+
 header = "my header"
 
 if rank == 0
@@ -38,11 +40,15 @@ byte_offset = MPI.File.get_byte_offset(fh, offset)
 MPI.File.set_view!(fh, byte_offset, MPI.Datatype(Int64), MPI.Datatype(Int64))
 @test MPI.File.get_position_shared(fh) == 0
 
+MPI.Barrier(comm)
+
 MPI.File.write_ordered(fh, fill(Int64(rank), rank+1))
 @test MPI.File.get_position_shared(fh) == sum(1:sz)
 
 MPI.File.seek_shared(fh, 0)
 @test MPI.File.get_position_shared(fh) == 0
+
+MPI.Barrier(comm)
 
 buf = zeros(Int64, rank+1)
 MPI.File.read_ordered!(fh, buf)
@@ -54,6 +60,8 @@ MPI.Barrier(comm)
 MPI.File.set_view!(fh, 0, MPI.Datatype(UInt8), MPI.Datatype(UInt8))
 MPI.File.seek_shared(fh, 0)
 @test MPI.File.get_position_shared(fh) == 0
+
+MPI.Barrier(comm)
 
 if rank == sz-1
     buf = Array{UInt8}(undef, sizeof(header))
