### 2021-06-07 cd3648b
Update TagBot setup and reduce frequency of CompatHelper to once per day
[skip ci]
