### 2021-06-07 a292068
Update TagBot setup and reduce frequency of CompatHelper to once per day (#475)
