### 2021-06-30 ceb39ff
simplify test/wait
