+++ test/test_sendrecv.jl
@@ -29,7 +29,7 @@ fill!(recv_mesg_expected, Float64(src))
 rreq = MPI.Irecv!(recv_mesg, src,  src+32, comm)
 sreq = MPI.Isend(send_mesg, dst, rank+32, comm)
 
-stats = MPI.Waitall!([sreq, rreq])
+stats = MPI.Waitall([sreq, rreq])
 @test rreq isa MPI.Request
 @test sreq isa MPI.Request
 @test MPI.Get_source(stats[2]) == src
@@ -67,15 +67,14 @@ rreq = MPI.Irecv!(recv_mesg, src,  src+32, comm)
 sreq = MPI.Isend(send_mesg, dst, rank+32, comm)
 
 req_arr = [sreq,rreq]
-(inds, stats) = MPI.Waitsome!(req_arr)
+inds = MPI.Waitsome(req_arr)
 for i in inds
-    (done, status) = MPI.Test!( req_arr[i] )
-    @test done
+    @test MPI.Test(req_arr[i])
 end
 
 rreq = MPI.Irecv!(recv_mesg, src,  src+32, comm)
 MPI.Cancel!(rreq)
-MPI.Wait!(rreq)
+MPI.Wai(rreq)
 @test rreq.buffer == nothing
 
 GC.gc()
@@ -132,7 +131,7 @@ MPI.Sendrecv!(a, dest_rank, 2,
 
 @test b == [(comm_rank+1) % comm_size, (comm_rank+1) % comm_size, (comm_rank+1) % comm_size]
 
-@test MPI.Waitall!(MPI.Request[]) == MPI.Status[]
+@test MPI.Waitall(MPI.Request[]) == MPI.Status[]
 
 MPI.Finalize()
 # @test MPI.Finalized()

[…]
