### 2021-07-02 #68 @crstnbr
@simonbyrne, can you please archive the current state of MPI.jl using Zenodo and post the associated DOI here (see https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance).
