### 2021-07-04 4aa2bc6
Remove `Pkg` from test environment
`Pkg` is already present in the main environment and this causes a warning when
running the tests in Julia v1.8:

```
(jl_H7V8sZ) pkg> test MPI
     Testing MPI
┌ Warning: Could not use exact versions of packages in manifest, re-resolving
└ @ Pkg.Operations ~/repo/julia/usr/share/julia/stdlib/v1.8/Pkg/src/Operations.jl:1497
┌ Warning: Dependency graph not a DAG, linearizing anyway
└ @ Pkg.Operations ~/repo/julia/usr/share/julia/stdlib/v1.8/Pkg/src/Operations.jl:840
┌ Warning: Dependency graph not a DAG, linearizing anyway
└ @ Pkg.Operations ~/repo/julia/usr/s
[…]
