### 2021-07-05 68545ed
[mpiexecjl] Accept `JULIA_BINDIR` as environment variable (#481)
[`JULIA_BINDIR`](https://docs.julialang.org/en/v1/manual/environment-variables/#JULIA_BINDIR)
is a standard Julia environment variable pointing to the directory where the
`julia` binary is present.
