### 2021-07-07 #68 @crstnbr
@simonbyrne There seems to be a missing DOI for "Open MPI: Goals, Concept, and Design of a Next Generation MPI Implementation" (see above).
