+++ paper/paper.bib
@@ -44,7 +44,8 @@
   year         = 2004,
   address      = {Budapest, Hungary},
   pages        = {97--104},
-  month        = {September}
+  month        = {September},
+  doi = {10.1007/978-3-540-30218-6_19}
 } 
 
 @article{clima,
