### 2021-07-07 #68 @crstnbr
@vchuravy: The proof looks good to me. Please go ahead and `accept deposit=true from branch paper`!
